+++ drools-platform-manager/drools-platform-persistence/src/main/java/org/chtijbug/drools/platform/persistence/pojo/DroolsResource.java
@@ -45,9 +45,9 @@ public class DroolsResource {
     private Date startDate;
     @Column
     private Date endDate;
-    private Integer startEventID;
+    private Long startEventID;
 
-    private Integer stopEventID;
+    private Long stopEventID;
 
 
     public DroolsResource() {
@@ -117,19 +117,19 @@ public class DroolsResource {
         this.guvnor_packageVersion = guvnor_packageVersion;
     }
 
-    public Integer getStartEventID() {
+    public Long getStartEventID() {
         return startEventID;
     }
 
-    public void setStartEventID(Integer startEventID) {
+    public void setStartEventID(Long startEventID) {
         this.startEventID = startEventID;
     }
 
-    public Integer getStopEventID() {
+    public Long getStopEventID() {
         return stopEventID;
     }
 
-    public void setStopEventID(Integer stopEventID) {
+    public void setStopEventID(Long stopEventID) {
         this.stopEventID = stopEventID;
     }
